# valentin.poli-jws-2024/src/main/java/fr/epita/assistant/jws/presentation/rest/request/CreateGameRequestDTO.java
package fr.epita.assistant.jws.presentation.rest.request;

import fr.epita.assistant.jws.utils.GameState;
import lombok.AllArgsConstructor;
import lombok.NoArgsConstructor;

@AllArgsConstructor @NoArgsConstructor
public class CreateGameRequestDTO {
    public String name;
}